+++ src/plugins/taxonomy/index.js
@@ -12,7 +12,7 @@ import { serviceIds } from './constants';
 
 export default class TaxonomyPlugin extends Plugin {
   constructor() {
-    super('triniti', 'taxonomy', '0.2.3');
+    super('triniti', 'taxonomy', '0.2.4');
   }
 
   configure(app, bottle) {