+++ src/plugins/utils/index.js
@@ -4,7 +4,7 @@ import reducer from './reducers';
 
 export default class UtilsPlugin extends Plugin {
   constructor() {
-    super('triniti', 'utils', '0.2.3');
+    super('triniti', 'utils', '0.2.4');
   }
 
   configure() {